+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   Button to download the perfs of multiple charts at once (#312)
+-   Buttons to download the tested models in a test task drawer (#319)
 
 ### Fixed